+++ resources/views/components/accordion.blade.php
@@ -9,9 +9,8 @@
 @endphp
 
 <div
-    class="{{ implode(' ', $classes) }}"
     x-data="{ openItems: [], multiple: {{ $multiple ? 'true' : 'false' }}, toggle(name) { if (this.openItems.includes(name)) { this.openItems = this.openItems.filter(i => i !== name); } else { if (!this.multiple) this.openItems = []; this.openItems.push(name); } } }"
-    {{ $attributes }}
+    {{ $attributes->class($classes) }}
 >
     {{ $slot }}
 </div>