+++ resources/views/components/dropdown/separator.blade.php
@@ -1,3 +1,2 @@
 @props([])
-
-<div class="aura-dropdown-separator h-px my-1.5 bg-aura-surface-200" role="separator" {{ $attributes }}></div>
+<div role="separator" {{ $attributes->class(['aura-dropdown-separator h-px my-1.5 bg-aura-surface-200']) }}></div>