+++ resources/views/components/tabs/tab.blade.php
@@ -3,11 +3,11 @@
 ])
 
 <div
-    class="aura-tabs-panel"
+   
     role="tabpanel"
     x-show="activeTab === {{ Js::from($name) }}"
     x-cloak
-    {{ $attributes }}
+    {{ $attributes->class(['aura-tabs-panel']) }}
 >
     {{ $slot }}
 </div>