+++ resources/views/components/toggle.blade.php
@@ -22,9 +22,8 @@
 <label class="aura-toggle-wrapper inline-flex items-center gap-2.5 cursor-pointer select-none relative" @if($disabled) aria-disabled="true" @endif>
     <input
         type="checkbox"
-        class="aura-toggle-input"
         @if($disabled) disabled @endif
-        {{ $attributes }}
+        {{ $attributes->class(['aura-toggle-input']) }}
     />
     <span class="{{ $trackClasses }} aura-toggle-{{ $color }} relative bg-aura-surface-300 border border-aura-surface-500 rounded-aura-full aura-transition-slow shrink-0">
         <span class="aura-toggle-knob {{ $knobClasses }} absolute top-px left-px bg-white rounded-full shadow-[0_1px_3px_rgba(0,0,0,0.15),0_1px_2px_rgba(0,0,0,0.1)] aura-transition-slow"></span>